[#85] Fixed behavior of OUTI,OTIR,OUTD,OTDR instructions
+ test suite refactoring

+++ plugins/cpu/8080-cpu/src/test/java/net/sf/emustudio/intel8080/impl/suite/FlagsBuilderImpl.java
@@ -12,7 +12,7 @@ public class FlagsBuilderImpl<T extends Number> extends FlagsBuilder<T, FlagsBui
 
     public FlagsBuilderImpl sign() {
         evaluators.add((context, result) -> {
-            if ((byte)result < 0) {
+            if (result.byteValue() < 0) {
                 expectedFlags |= FLAG_S;
             } else {
                 expectedNotFlags |= FLAG_S;
@@ -23,7 +23,7 @@ public class FlagsBuilderImpl<T extends Number> extends FlagsBuilder<T, FlagsBui
 
     public FlagsBuilderImpl zero() {
         evaluators.add((context, result) -> {
-            if ((byte)result == 0) {
+            if (result.byteValue() == 0) {
                 expectedFlags |= FLAG_Z;
             } else {
                 expectedNotFlags |= FLAG_Z;
@@ -35,7 +35,7 @@ public class FlagsBuilderImpl<T extends Number> extends FlagsBuilder<T, FlagsBui
     public FlagsBuilderImpl parity() {
         evaluators.add((context, result) -> {
             int numberOf1 = 0;
-            int intResult = result & 0xFF;
+            int intResult = result.intValue() & 0xFF;
 
             for (int i = 0; i < 8; i++) {
                 if ((intResult & 1) == 1) {
@@ -55,7 +55,7 @@ public class FlagsBuilderImpl<T extends Number> extends FlagsBuilder<T, FlagsBui
 
     public FlagsBuilderImpl carry15() {
         evaluators.add((context, result) -> {
-            if ((result & 0x10000) == 0x10000) {
+            if ((result.intValue() & 0x10000) == 0x10000) {
                 expectedFlags |= FLAG_C;
             } else {
                 expectedNotFlags |= FLAG_C;
@@ -66,7 +66,7 @@ public class FlagsBuilderImpl<T extends Number> extends FlagsBuilder<T, FlagsBui
 
     public FlagsBuilderImpl carry() {
         evaluators.add((context, result) -> {
-            if ((result & 0x100) == 0x100) {
+            if ((result.intValue() & 0x100) == 0x100) {
                 expectedFlags |= FLAG_C;
             } else {
                 expectedNotFlags |= FLAG_C;
@@ -117,7 +117,7 @@ public class FlagsBuilderImpl<T extends Number> extends FlagsBuilder<T, FlagsBui
     public FlagsBuilderImpl auxCarry() {
         evaluators.add((context, result) -> {
             int firstInt = context.first.intValue();
-            byte diff = (byte)((result - firstInt) & 0xFF);
+            byte diff = (byte)((result.intValue() - firstInt) & 0xFF);
 
             if (isAuxCarry(context.first.intValue(), diff)) {
                 expectedFlags |= FLAG_AC;
